+++ examples/src/polytype.mli
@@ -5,3 +5,7 @@ type 'a llist =
 type box = Box : 'a -> box
 
 type (_, 'a) gadt = Gadt : int * 'a -> (int, 'a) gadt
+
+type _ m =
+  | M1 : box m
+  | M2 : (int, box llist) gadt m